requirements.txt
praw==7.7.1
google-generativeai==0.8.3
python-dotenv==1.0.1
textblob==0.18.0
nltk==3.9.1